Adding ability to skip dependency installs

+++ centos7/start.sh
@@ -8,16 +8,17 @@ playbook=''
 galaxyfile='requirements.yml'
 pypifile='requirements.txt'
 pkgfile='system_packages.txt'
-inventory='localhost,'
+skip_all=0
 cmd="ansible-playbook"
 
-USAGE="""$0 [-p test.yml] [-g requirements.yml] [-r requirements.txt] [-s system_packages.txt] [-*] [-h]
+USAGE="""$0 [-p test.yml] [-g requirements.yml] [-r requirements.txt] [-s system_packages.txt] [-x] [-*] [-h]
   Installs pre-reqs and runs an Ansible playbook.
 
   -p    Path to Ansible playbook (default: test.yml > local.yml > playbook.yml > site.yml)
   -g    Path to Ansible Galaxy requirements file (default: requirements.yml)
   -r    Path to PyPI/pip requirements file (default: requirements.txt)
   -s    Path to a list of system packages to install, one per line. (default: system_packages.txt)
+  -x    Skip all dependency installs.
   -*    Any option supported by ansible-playbook (eg: -e SOMEVAR=someval -i /path/to/inventory)
   -v    Enable debug messages
   -h    Show this help message
@@ -35,6 +36,8 @@ while test $# -gt 0; do
       pypifile="${2}" #; echo "pypifile=${pypifile}"
     elif [ "$1" == "-s" ]; then
       pkgfile="${2}" #; echo "pkgfile=${pkgfile}"
+    elif [ "$1" == "-x" ]; then
+      skip_all=1
     elif [ "$1" == "-h" ]; then
       echo $USAGE; exit 0
     else
@@ -45,20 +48,20 @@ while test $# -gt 0; do
 done
 
 # Install system packages
-if [ -f "${pkgfile}" ]; then
+if [ -f "${pkgfile}" ] && [[ $skip_all -eq 0 ]]; then
   echo -e "\n### Installing system packages..."
   pkgs=$(<$pkgfile)
   yum makecache fast; yum install -y $pkgs
 fi
 
 # Install ansible-galaxy requirements
-if [ -f "${galaxyfile}" ]; then
+if [ -f "${galaxyfile}" ] && [[ $skip_all -eq 0 ]]; then
   echo -e "\n### Installing pre-reqs from Ansible Galaxy..."
   ansible-galaxy install -r "${galaxyfile}"
 fi
 
 # Install Python requirements
-if [ -f "${pypifile}" ]; then
+if [ -f "${pypifile}" ] && [[ $skip_all -eq 0 ]]; then
   echo -e "\n### Installing pre-reqs from PyPI..."
   pip install -r "${pypifile}"
 fi
